requirements.txt
fastapi==0.104.1
sqlalchemy==2.0.23
requests==2.31.0
beautifulsoup4==4.12.2
celery==5.3.4
redis==5.0.1
openai==1.3.8
sentence-transformers==5.1.0
pandas==2.1.3
uvicorn==0.24.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
alembic==1.12.1
python-multipart==0.0.6
pydantic==2.5.0
schedule==1.2.2
bcrypt==4.0.1
email-validator==2.3.0